+++ source/domain/ball.h
@@ -4,7 +4,7 @@
 #include <memory>
 #include <experimental/propagate_const>
 
-#include "utils/vector2i.h"
+#include "utils/vector2f.h"
 
 namespace domain {
 
@@ -12,21 +12,21 @@ class Ball
 {
 public:
     Ball();
-    Ball(const utils::Vector2i& position);
+    Ball(const utils::Vector2f& position);
     ~Ball();
 
-    void setPosition(const utils::Vector2i& newPosition);
-    const utils::Vector2i& position() const;
+    void setPosition(const utils::Vector2f& newPosition);
+    const utils::Vector2f& position() const;
 
-    void setVelocity(const utils::Vector2i& newVelocity);
-    const utils::Vector2i& velocity() const;
+    void setVelocity(const utils::Vector2f& newVelocity);
+    const utils::Vector2f& velocity() const;
 
     void setFixed(bool fixed);
     bool isFixed() const;
 
     constexpr int radius();
 
-    void applyForce(const utils::Vector2i& force);
+    void applyForce(const utils::Vector2f& force);
     void makeStep();
 
 private:
